add LOGIC

+++ unpackota.sh
@@ -1,8 +1,14 @@
-mv latest.tar latest.ota
-mv latest.ota latest.tar
-echo opening ota
-tar -xf latest.tar
-rm opened/apq8009-* opened/manifest.*
-mv manifest.* opened
-mv apq8009-* opened
-echo done
+#!/usr/bin/env bash
+
+if [ "$1" == "" ] then
+  echo no ota provided.
+else
+  echo opening ota
+  tar -xf $1
+  if [ -f "opened/apq8009*" ] then
+    rm opened/apq8009-* opened/manifest.*
+  fi
+  mv manifest.* opened
+  mv apq8009-* opened
+  echo done
+fi
